requirements.txt
# ============================================
# Core Django Framework
# ============================================
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# ============================================
# Security & IP Tracking (from alx-backend-security)
# ============================================
django-ipware>=5.0.0
django-ratelimit>=4.1.0
geoip2>=4.8.0

# ============================================
# Caching & Redis (from alx-backend-caching_property_listings)
# ============================================
django-redis>=5.4.0
redis>=5.0.0

# ============================================
# GraphQL (from alx-backend-graphql_crm)
# ============================================
graphene-django>=3.2.0
django-filter>=25.0
django-crontab>=0.7.1
django-celery-beat>=2.5.0
graphene>=3.0.0

# ============================================
# Celery & Task Queue
# ============================================
celery>=5.3.0
kombu>=5.3.0
amqp>=5.2.0  # For RabbitMQ support

# ============================================
# API Documentation (Swagger)
# ============================================
drf-yasg>=1.21.0

# ============================================
# Database
# ============================================
psycopg2-binary>=2.9.0
mysqlclient>=2.2.0  # For MySQL support (if needed)

# ============================================
# Environment Variables
# ============================================
django-environ>=0.11.0
python-decouple>=3.8
dj-database-url>=2.1.0

# ============================================
# Data Science & Analytics (from alx-backend-security)
# ============================================
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# ============================================
# Authentication & JWT (from alx-backend-python)
# ============================================
djangorestframework-simplejwt>=5.3.0
drf-nested-routers>=0.93.5

# ============================================
# Production Server
# ============================================
gunicorn>=21.2.0
whitenoise>=6.6.0  # For static file serving

# ============================================
# Development Tools
# ============================================
python-dotenv>=1.0.0

# ============================================
# Testing (Optional)
# ============================================
pytest>=7.4.0
pytest-django>=4.7.0
coverage>=7.3.0
